+++ android/src/oldarch/AndroidWidgetSpec.java
@@ -20,5 +20,7 @@ abstract class AndroidWidgetSpec extends ReactContextBaseJavaModule {
 
     public abstract void getWidgetInfo(String widgetName, Promise promise);
 
+    public abstract void requestPinWidget(String widgetName, Promise promise);
+
     public abstract void finishWidgetConfiguration(double widgetId, String result);
 }