+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `requestPinWidget` to show the launcher's native add-widget prompt on supported Android launchers
+
 ## [0.21.0] - 2026-07-11
 
 ### Added